adjust project architecture
